Fixes for 1.6 release.

+++ EsentInterop/jet_unicodeindex.cs
@@ -121,7 +121,7 @@ namespace Microsoft.Isam.Esent.Interop
             var native = new NATIVE_UNICODEINDEX
             {
                 lcid = (uint)this.lcid,
-                dwMapFlags = (uint)this.dwMapFlags,
+                dwMapFlags = this.dwMapFlags,
             };
             return native;
         }
